First checker commit of exercise 8.16.
This is the first commit of checker in exercise 8.16. It is not finished,
because it was written after two days' work. In the past two days I became a
code-machine, woke up at 9:00, arrived at work at 10:00, coded till 22:00, and
fell into bed once arriving at home at 23:00. This is another fresh proof of
'996, ICU' in China. From the perspective of a programmer, this is nothing but
life-consuming (or 'passionate' for those who long for changing the
world?). From the perspective of a big boss, this is what you deserve as a
worker. From the perspective of an anthropologist, this is a deep reflection of
the human race worthy of study. But for me, I just want to leave it,
forever. How ironic, to program I have to quit the job as a programmer!